vendr/requirements.txt
bcrypt==3.1.3
boto3==1.4.5
botocore==1.5.92
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
Django==1.10.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-model-utils==3.0.0
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
docutils==0.14
futures==3.1.1
gunicorn==19.7.1
gyp==0.1
idna==2.5
jmespath==0.9.3
lxml==3.8.0
oauth2-provider==0.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
praw==5.0.1
prawcore==0.11.0
psycopg2==2.7.3
pycparser==2.18
python-dateutil==2.6.1
redis==2.10.5
requests==2.18.3
s3transfer==0.1.10
six==1.10.0
update-checker==0.16
urllib3==1.22
